requirements.txt
fastapi==0.92.0
pip==23.0.1
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytest==7.2.1
sqlalchemy==2.0.4

uvicorn~=0.20.0